[in_app_purchase] Android example using in_app_purchase_android package (#3861)
Adds the example app for the in_app_purchase_android platform implementation.

Adds the Android implementation for issue flutter/flutter#81695
NOTE: this PR builds on top of the "[in_app_purchase] Federated Android implementation" pull request. If have split in into a separate PR so it would be easier to review.