Remove -Werror from deprecated plugin Android builds (#4466)
Now that the default Flutter app template uses a newer compile SDK on Android, the build_all test (which generates an application) is subject to new out-of-band deprecation warnings. For plugins with `-Werror` specified, these are errors, which break CI.

To prevent this, remove `-Werror` from the Android builds of all the deprecated plugins that used it.